Move genric implementations into corresponding files instead
