+++ data/binding/trees.go
@@ -416,3 +416,284 @@ func parentIDFor(id string, ids map[string][]string) string {
 
 	return ""
 }
+
+func newTree[T any](comparator func(T, T) bool) *boundTree[T] {
+	t := &boundTree[T]{val: &map[string]T{}, comparator: comparator}
+	t.ids = make(map[string][]string)
+	t.items = make(map[string]DataItem)
+	return t
+}
+
+func newTreeComparable[T bool | float64 | int | rune | string]() *boundTree[T] {
+	return newTree(func(t1, t2 T) bool { return t1 == t2 })
+}
+
+func bindTree[T any](ids *map[string][]string, v *map[string]T, comparator func(T, T) bool) *boundTree[T] {
+	if v == nil {
+		return newTree[T](comparator)
+	}
+
+	t := &boundTree[T]{val: v, updateExternal: true, comparator: comparator}
+	t.ids = make(map[string][]string)
+	t.items = make(map[string]DataItem)
+
+	for parent, children := range *ids {
+		for _, leaf := range children {
+			t.appendItem(bindTreeItem(v, leaf, t.updateExternal, t.comparator), leaf, parent)
+		}
+	}
+
+	return t
+}
+
+func bindTreeComparable[T bool | float64 | int | rune | string](ids *map[string][]string, v *map[string]T) *boundTree[T] {
+	return bindTree(ids, v, func(t1, t2 T) bool { return t1 == t2 })
+}
+
+type boundTree[T any] struct {
+	treeBase
+
+	comparator     func(T, T) bool
+	val            *map[string]T
+	updateExternal bool
+}
+
+func (t *boundTree[T]) Append(parent, id string, val T) error {
+	t.lock.Lock()
+	ids, ok := t.ids[parent]
+	if !ok {
+		ids = make([]string, 0)
+	}
+
+	t.ids[parent] = append(ids, id)
+	v := *t.val
+	v[id] = val
+
+	trigger, err := t.doReload()
+	t.lock.Unlock()
+
+	if trigger {
+		t.trigger()
+	}
+
+	return err
+}
+
+func (t *boundTree[T]) Get() (map[string][]string, map[string]T, error) {
+	t.lock.RLock()
+	defer t.lock.RUnlock()
+
+	return t.ids, *t.val, nil
+}
+
+func (t *boundTree[T]) GetValue(id string) (T, error) {
+	t.lock.RLock()
+	defer t.lock.RUnlock()
+
+	if item, ok := (*t.val)[id]; ok {
+		return item, nil
+	}
+
+	return *new(T), errOutOfBounds
+}
+
+func (t *boundTree[T]) Prepend(parent, id string, val T) error {
+	t.lock.Lock()
+	ids, ok := t.ids[parent]
+	if !ok {
+		ids = make([]string, 0)
+	}
+
+	t.ids[parent] = append([]string{id}, ids...)
+	v := *t.val
+	v[id] = val
+
+	trigger, err := t.doReload()
+	t.lock.Unlock()
+
+	if trigger {
+		t.trigger()
+	}
+
+	return err
+}
+
+func (t *boundTree[T]) Remove(id string) error {
+	t.lock.Lock()
+	t.removeChildren(id)
+	delete(t.ids, id)
+	v := *t.val
+	delete(v, id)
+
+	trigger, err := t.doReload()
+	t.lock.Unlock()
+
+	if trigger {
+		t.trigger()
+	}
+
+	return err
+}
+
+func (t *boundTree[T]) removeChildren(id string) {
+	for _, cid := range t.ids[id] {
+		t.removeChildren(cid)
+
+		delete(t.ids, cid)
+		v := *t.val
+		delete(v, cid)
+	}
+}
+
+func (t *boundTree[T]) Reload() error {
+	t.lock.Lock()
+	trigger, err := t.doReload()
+	t.lock.Unlock()
+
+	if trigger {
+		t.trigger()
+	}
+
+	return err
+}
+
+func (t *boundTree[T]) Set(ids map[string][]string, v map[string]T) error {
+	t.lock.Lock()
+	t.ids = ids
+	*t.val = v
+
+	trigger, err := t.doReload()
+	t.lock.Unlock()
+
+	if trigger {
+		t.trigger()
+	}
+
+	return err
+}
+
+func (t *boundTree[T]) doReload() (fire bool, retErr error) {
+	updated := []string{}
+	for id := range *t.val {
+		found := false
+		for child := range t.items {
+			if child == id { // update existing
+				updated = append(updated, id)
+				found = true
+				break
+			}
+		}
+		if found {
+			continue
+		}
+
+		// append new
+		t.appendItem(bindTreeItem(t.val, id, t.updateExternal, t.comparator), id, parentIDFor(id, t.ids))
+		updated = append(updated, id)
+		fire = true
+	}
+
+	for id := range t.items {
+		remove := true
+		for _, done := range updated {
+			if done == id {
+				remove = false
+				break
+			}
+		}
+
+		if remove { // remove item no longer present
+			fire = true
+			t.deleteItem(id, parentIDFor(id, t.ids))
+		}
+	}
+
+	for id, item := range t.items {
+		var err error
+		if t.updateExternal {
+			err = item.(*boundExternalTreeItem[T]).setIfChanged((*t.val)[id])
+		} else {
+			err = item.(*boundTreeItem[T]).doSet((*t.val)[id])
+		}
+		if err != nil {
+			retErr = err
+		}
+	}
+	return
+}
+
+func (t *boundTree[T]) SetValue(id string, v T) error {
+	t.lock.Lock()
+	(*t.val)[id] = v
+	t.lock.Unlock()
+
+	item, err := t.GetItem(id)
+	if err != nil {
+		return err
+	}
+	return item.(bindableItem[T]).Set(v)
+}
+
+func bindTreeItem[T any](v *map[string]T, id string, external bool, comparator func(T, T) bool) bindableItem[T] {
+	if external {
+		ret := &boundExternalTreeItem[T]{old: (*v)[id], comparator: comparator}
+		ret.val = v
+		ret.id = id
+		return ret
+	}
+
+	return &boundTreeItem[T]{id: id, val: v}
+}
+
+type boundTreeItem[T any] struct {
+	base
+
+	val *map[string]T
+	id  string
+}
+
+func (t *boundTreeItem[T]) Get() (T, error) {
+	t.lock.Lock()
+	defer t.lock.Unlock()
+
+	v := *t.val
+	if item, ok := v[t.id]; ok {
+		return item, nil
+	}
+
+	return *new(T), errOutOfBounds
+}
+
+func (t *boundTreeItem[T]) Set(val T) error {
+	return t.doSet(val)
+}
+
+func (t *boundTreeItem[T]) doSet(val T) error {
+	t.lock.Lock()
+	(*t.val)[t.id] = val
+	t.lock.Unlock()
+
+	t.trigger()
+	return nil
+}
+
+type boundExternalTreeItem[T any] struct {
+	boundTreeItem[T]
+
+	comparator func(T, T) bool
+	old        T
+}
+
+func (t *boundExternalTreeItem[T]) setIfChanged(val T) error {
+	t.lock.Lock()
+	if t.comparator(val, t.old) {
+		t.lock.Unlock()
+		return nil
+	}
+	(*t.val)[t.id] = val
+	t.old = val
+	t.lock.Unlock()
+
+	t.trigger()
+	return nil
+}
